Add BooleanArray::from_unary and BooleanArray::from_binary
